Refactored CSS and SASS
SASS organized by section. CSS is now organized top to bottom from how they appear on the page. All colors replaced with variables to lower the number of places needed to change the colors. Dark mode/ light mode switcher totally removed and replaced with the Styled Components package. Color variables are stored in Themes.jsx. Theme colors are applied in the GlobalStyles.jsx. The advantage of this is the colors can be changed in one place. Styled components was setup to remember the users theme preference between page refreshes. When the user Changes the theme there is also a fade transition now.