All px units changed to rem and other changes
Fonts changed to rem units because it allows you to change the zoom level of the whole page at once. Some users have a zoomed in browser, this will make the text resize correctly. The theme switcher button now highlights the currently active language flag. This is less confusing this way. Previously the current language flag was blacked out and the disabled one was highlighted. Removed more CSS rules that either did not work or did not have a purpose.